updates contract artifacts
