+++ packages/web/health-check-sha.json
@@ -1,1 +1,1 @@
-{ "artifactsSha": "273f70f22", "commitSha": "0ddb218" }
+{ "artifactsSha": "e6899d64e", "commitSha": "d844a57" }
